# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AutoMind Smart Car Discovery API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterBody {
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface UserProfile {
  id: number;
  name: string;
  email: string;
}

export interface AuthResponse {
  token: string;
  user: UserProfile;
}

export interface Car {
  car_id: number;
  brand: string;
  model: string;
  variant_name: string;
  year: number;
  body_type: string;
  fuel_type: string;
  transmission: string;
  price_lakh: number;
  mileage_kmpl?: number;
  engine_cc?: number;
  seating_capacity?: number;
  service_cost_inr?: number;
  airbags?: number;
  abs?: string;
  rear_camera?: string;
  touchscreen?: string;
  cruise_control?: string;
  sunroof?: string;
  safety_rating?: number;
  brand_image_score?: number;
  sentiment_score?: number;
  headlight_inr?: number;
  backlight_inr?: number;
  front_bumper_inr?: number;
  rear_bumper_inr?: number;
  side_mirror_inr?: number;
  score?: number;
  image_url?: string;
}

export type CarDetail = Car;

export interface CarsListResponse {
  cars: Car[];
  total: number;
  page: number;
  limit: number;
  corrected_query?: string;
}

export interface BrandsResponse {
  brands: string[];
}

export type CarStatsBodyTypeCounts = { [key: string]: number };

export type CarStatsFuelTypeCounts = { [key: string]: number };

export type CarStatsPriceRange = {
  min: number;
  max: number;
};

export interface CarStats {
  total_cars: number;
  total_brands: number;
  avg_price_lakh: number;
  avg_mileage: number;
  body_type_counts: CarStatsBodyTypeCounts;
  fuel_type_counts: CarStatsFuelTypeCounts;
  price_range: CarStatsPriceRange;
}

export type RecommendBodyPriority =
  (typeof RecommendBodyPriority)[keyof typeof RecommendBodyPriority];

export const RecommendBodyPriority = {
  mileage: "mileage",
  safety: "safety",
  value: "value",
  comfort: "comfort",
} as const;

export interface RecommendBody {
  budget_max?: number;
  budget_min?: number;
  fuel_type?: string;
  transmission?: string;
  body_type?: string;
  seating_capacity?: number;
  priority?: RecommendBodyPriority;
}

export interface RecommendResponse {
  recommendations: Car[];
  total: number;
}

export interface CompareBody {
  /**
   * @minItems 2
   * @maxItems 4
   */
  car_ids: number[];
}

export type CompareResponseWinner = {
  car_id?: number;
  reason?: string;
};

export interface CompareResponse {
  cars: Car[];
  winner?: CompareResponseWinner;
}

export interface QueryProcessBody {
  query: string;
}

export type QueryProcessResponseFilters = {
  brand?: string;
  model?: string;
  body_type?: string;
  fuel_type?: string;
  transmission?: string;
  budget_max?: number;
  budget_min?: number;
  seating_capacity?: number;
};

export interface QueryProcessResponse {
  original_query: string;
  corrected_query: string;
  filters: QueryProcessResponseFilters;
  cars: Car[];
  total: number;
  message?: string;
}

export interface CarImageResponse {
  car_id: number;
  image_url: string;
  fallback: boolean;
}

export interface InsightsResponse {
  car_id: number;
  brand: string;
  model: string;
  why_recommended: string;
  strengths: string[];
  weaknesses: string[];
  best_for: string;
  value_score: number;
}

export type MarketOverviewTopBrandsItem = {
  brand: string;
  count: number;
  avg_score: number;
};

export type MarketOverviewPopularBodyTypesItem = {
  body_type: string;
  count: number;
};

export type MarketOverviewBudgetSegmentsItem = {
  segment: string;
  range: string;
  count: number;
};

export interface MarketOverview {
  top_brands: MarketOverviewTopBrandsItem[];
  popular_body_types: MarketOverviewPopularBodyTypesItem[];
  budget_segments: MarketOverviewBudgetSegmentsItem[];
  best_value_cars: Car[];
}

export type GetCarsParams = {
  brand?: string;
  model?: string;
  body_type?: string;
  fuel_type?: string;
  transmission?: string;
  min_price?: number;
  max_price?: number;
  seating_capacity?: number;
  search?: string;
  page?: number;
  limit?: number;
};