# lib/api-zod/src/generated/types/carDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AutoMind Smart Car Discovery API
 * OpenAPI spec version: 0.1.0
 */
import type { Car } from "./car";

export type CarDetail = Car;