# lib/api-zod/src/generated/types/compareBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AutoMind Smart Car Discovery API
 * OpenAPI spec version: 0.1.0
 */

export interface CompareBody {
  /**
   * @minItems 2
   * @maxItems 4
   */
  car_ids: number[];
}