# lib/api-zod/src/generated/types/marketOverview.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AutoMind Smart Car Discovery API
 * OpenAPI spec version: 0.1.0
 */
import type { Car } from "./car";
import type { MarketOverviewBudgetSegmentsItem } from "./marketOverviewBudgetSegmentsItem";
import type { MarketOverviewPopularBodyTypesItem } from "./marketOverviewPopularBodyTypesItem";
import type { MarketOverviewTopBrandsItem } from "./marketOverviewTopBrandsItem";

export interface MarketOverview {
  top_brands: MarketOverviewTopBrandsItem[];
  popular_body_types: MarketOverviewPopularBodyTypesItem[];
  budget_segments: MarketOverviewBudgetSegmentsItem[];
  best_value_cars: Car[];
}